component分割 #11
